[export] Fix run_decomps to work with fake mode
